Add research documentation to README
Insert links to web pages with information used in the creation of
iPhoto2Web. Insert meanings of integer values used in
SqFileImage.imageType and SqFileInfo.format. The latter were determined
through primary research, and the former were confirmed. Demonstrate how
to decode the various dates stored for photos and events.